+++ src/sdc11073/provider/providerimpl.py
@@ -11,10 +11,10 @@ from sdc11073 import loghelper
 from sdc11073 import observableproperties as properties
 from sdc11073.dispatch import (
     DispatchKey,
-    RequestDispatcher,
     MessageConverterMiddleware,
     PathElementRegistry,
     RequestData,
+    RequestDispatcher,
 )
 from sdc11073.exceptions import ApiUsageError
 from sdc11073.httpserver import compression
@@ -25,24 +25,25 @@ from sdc11073.xml_types import mex_types
 from sdc11073.xml_types.addressing_types import EndpointReferenceType
 from sdc11073.xml_types.dpws_types import HostServiceType, ThisDeviceType, ThisModelType
 from sdc11073.xml_types.wsd_types import ProbeMatchesType, ProbeMatchType
+from sdc11073.roles.protocols import ProductProtocol, WaveformProviderProtocol # import here for code cov. :(
 
 from .components import default_sdc_provider_components
 from .periodicreports import PeriodicReportsHandler, PeriodicReportsNullHandler
-from .waveforms import WaveformSender
 
 if TYPE_CHECKING:
     from enum import Enum
     from sdc11073.location import SdcLocation
     from sdc11073.mdib.providermdib import ProviderMdib
+    from sdc11073.mdib.transactions import TransactionProcessor
+    from sdc11073.mdib.statecontainers import AbstractStateProtocol
+    from sdc11073.provider.porttypes.localizationservice import LocalizationStorage
     from sdc11073.pysoap.msgfactory import CreatedMessage
     from sdc11073.pysoap.soapenvelope import ReceivedSoapMessage
     from sdc11073.xml_types.msg_types import AbstractSet
     from sdc11073.xml_types.wsd_types import ScopesType
-    from sdc11073.provider.porttypes.localizationservice import LocalizationStorage
-    from .operations import OperationDefinitionBase
 
     from .components import SdcProviderComponents
-
+    from .operations import OperationDefinitionBase
 
 class _PathElementDispatcher(PathElementRegistry):
     """Dispatch to one of the registered instances, based on path element.
@@ -145,7 +146,6 @@ class SdcProvider:
             self._urlschema = 'http'
 
         self.collect_rt_samples_period = 0.1  # in seconds
-        self._waveform_sender = None
         self.contextstates_in_getmdib = self.DEFAULT_CONTEXTSTATES_IN_GETMDIB  # can be overridden per instance
         # look for schemas added by services
         additional_schema_specs = []
@@ -193,9 +193,10 @@ class SdcProvider:
         self._soap_client_pool = SoapClientPool(self._mk_soap_client, log_prefix)
         self._sco_operations_registries = {}  # key is mds handle ?
         self._service_factory = None
-        self.product_roles_lookup = {}
+        self.product_lookup: dict[str, ProductProtocol] = {}  # one product per sco,  key is a sco handle
         self.hosted_services = None
         self._periodic_reports_handler = PeriodicReportsNullHandler()
+        self.waveform_provider: WaveformProviderProtocol | None = None
         self._setup_components()
         self.base_urls = []  # will be set after httpserver is started
         properties.bind(device_mdib_container, transaction=self._send_episodic_reports)
@@ -250,8 +251,12 @@ class SdcProvider:
             product_roles = self._components.role_provider_class(self._mdib,
                                                                  sco_operations_registry,
                                                                  self._log_prefix)
-            self.product_roles_lookup[sco_descr.Handle] = product_roles
+            self.product_lookup[sco_descr.Handle] = product_roles
             product_roles.init_operations()
+        if self._components.waveform_provider_class is not None:
+            self.waveform_provider = self._components.waveform_provider_class(self._mdib,
+                                                                              self._log_prefix)
+
         # product roles might have added descriptors, set source mds for all
         self._mdib.xtra.set_all_source_mds()
 
@@ -262,8 +267,8 @@ class SdcProvider:
             return self.hosted_services.localization_service.localization_storage
         return None
 
-    def _on_get_metadata(self, request_data):  # pylint: disable=unused-argument
-        self._logger.info('_on_get_metadata from {}', request_data.peer_name)
+    def _on_get_metadata(self, request_data: RequestData) -> CreatedMessage:
+        self._logger.info('_on_get_metadata from %s', request_data.peer_name)
         metadata = mex_types.Metadata()
         section = mex_types.ThisModelMetadataSection()
         section.MetadataReference = self.model
@@ -295,7 +300,7 @@ class SdcProvider:
                     needed_namespaces.append(e)
         return self.msg_factory.mk_reply_soap_message(request_data, metadata, needed_namespaces)
 
-    def _on_probe_request(self, request):
+    def _on_probe_request(self, request: RequestData) -> CreatedMessage:
         _nsm = self._mdib.nsmapper
         probe_matches = ProbeMatchesType()
         probe_match = ProbeMatchType()
@@ -308,8 +313,9 @@ class SdcProvider:
         response.p_msg.header_info_block.To = WSA_ANONYMOUS
         return response
 
-    def set_location(self, location: SdcLocation,
-                     validators=None,
+    def set_location(self,
+                     location: SdcLocation,
+                     validators: list | None = None,
                      publish_now: bool = True):
         """:param location: an SdcLocation instance
         :param validators: a list of pmtypes.InstanceIdentifier objects or None; in that case the defaultInstanceIdentifiers member is used
@@ -325,9 +331,7 @@ class SdcProvider:
             self.publish()
 
     def publish(self):
-        """Publish device on the network (sends HELLO message)
-        :return:
-        """
+        """Publish device on the network (sends HELLO message)."""
         scopes = self._components.scopes_factory(self._mdib)
         x_addrs = self.get_xaddrs()
         self._wsdiscovery.publish_service(self.epr_urn,
@@ -337,25 +341,27 @@ class SdcProvider:
 
     @property
     def mdib(self) -> ProviderMdib:
+        """Return mdib reference."""
         return self._mdib
 
     @property
-    def epr_urn(self):
-        # End Point Reference, e.g 'urn:uuid:8c26f673-fdbf-4380-b5ad-9e2454a65b6b'
+    def epr_urn(self) -> str:
+        """Return end point reference, e.g 'urn:uuid:8c26f673-fdbf-4380-b5ad-9e2454a65b6b'."""
         try:
             return self._epr.urn
         except AttributeError:
             return self._epr
 
     @property
-    def path_prefix(self):
-        # http path prefix of service e.g '8c26f673fdbf4380b5ad9e2454a65b6b'
+    def path_prefix(self) -> str:
+        """Return http path prefix of service e.g '8c26f673fdbf4380b5ad9e2454a65b6b'."""
         try:
             return self._epr.hex
         except AttributeError:
             return self._epr
 
-    def get_operation_by_handle(self, operation_handle):
+    def get_operation_by_handle(self, operation_handle: str) -> OperationDefinitionBase | None:
+        """Return OperationDefinitionBase for given handle or None if it does not exist."""
         for sco in self._sco_operations_registries.values():
             op = sco.get_operation_by_handle(operation_handle)
             if op is not None:
@@ -372,19 +378,13 @@ class SdcProvider:
             has_this_operation = sco.get_operation_by_handle(operation.handle) is not None
             if has_this_operation:
                 return sco.handle_operation_request(operation, request, operation_request, transaction_id)
-        self._logger.error('no sco has operation {}', operation.handle)
+        self._logger.error('no sco has operation %s', operation.handle)
         return self.mdib.data_model.msg_types.InvocationState.FAILED
 
-    def get_toplevel_sco_list(self) -> list:
-        pm_names = self._mdib.data_model.pm_names
-        mds_handles = [d.Handle for d in self._mdib.descriptions.NODETYPE.get(pm_names.MdsDescriptor, [])]
-        ret = []
-        for sco in self._sco_operations_registries.values():
-            if sco.sco_descriptor_container.parent_handle in mds_handles:
-                ret.append(sco)
-        return ret
-
-    def start_all(self, start_rtsample_loop=True, periodic_reports_interval=None, shared_http_server=None):
+    def start_all(self,
+                  start_rtsample_loop: bool =True,
+                  periodic_reports_interval: float | None = None,
+                  shared_http_server=None):
         """:param start_rtsample_loop: flag
         :param periodic_reports_interval: if provided, a value in seconds
         :param shared_http_server: if provided, use this http server, else device creates its own.
@@ -406,7 +406,7 @@ class SdcProvider:
 
     def _start_services(self, shared_http_server=None):
         """Start the services."""
-        self._logger.info('starting services, addr = {}', self._wsdiscovery.get_active_addresses())
+        self._logger.info('starting services, addr = %r', self._wsdiscovery.get_active_addresses())
         for sco in self._sco_operations_registries.values():
             sco.start_worker()
 
@@ -447,11 +447,11 @@ class SdcProvider:
                 SplitResult(self._urlschema, f'{addr}:{port}', self.path_prefix, query=None, fragment=None))
 
         for host_ip in host_ips:
-            self._logger.info('serving Services on {}:{}', host_ip, port)
+            self._logger.info('serving Services on %s:%d', host_ip, port)
         for subscriptions_manager in self._subscriptions_managers.values():
             subscriptions_manager.set_base_urls(self.base_urls)
 
-    def stop_all(self, send_subscription_end=True):
+    def stop_all(self, send_subscription_end: bool = True):
         self.stop_realtime_sample_loop()
         if self._periodic_reports_handler:
             self._periodic_reports_handler.stop()
@@ -462,24 +462,25 @@ class SdcProvider:
         try:
             self._wsdiscovery.clear_service(self.epr_urn)
         except KeyError:
-            self._logger.info('epr "{}" not known in self._wsdiscovery', self.epr_urn)
-        for role in self.product_roles_lookup.values():
+            self._logger.info('epr "%s" not known in self._wsdiscovery', self.epr_urn)
+        for role in self.product_lookup.values():
             role.stop()
         if self._is_internal_http_server and self._http_server is not None:
             self._http_server.stop()
         self._soap_client_pool.close_all()
 
     def start_rt_sample_loop(self):
-        if self._waveform_sender:
-            raise ApiUsageError(' realtime send loop already started')
-        self._waveform_sender = WaveformSender(self._mdib, self._logger, self.collect_rt_samples_period)
-        self._waveform_sender.start()
+        if self.waveform_provider is None:
+            raise ApiUsageError('no waveform provider configured.')
+        if self.waveform_provider.is_running:
+            raise ApiUsageError('realtime send loop already started')
+        self.waveform_provider.start()
 
     def stop_realtime_sample_loop(self):
-        if self._waveform_sender:
-            self._waveform_sender.stop()
+        if self.waveform_provider is not None and self.waveform_provider.is_running:
+            self.waveform_provider.stop()
 
-    def get_xaddrs(self):
+    def get_xaddrs(self) -> list[str]:
         addresses = self._wsdiscovery.get_active_addresses()  # these own IP addresses are currently used by discovery
         port = self._http_server.my_port
         xaddrs = []
@@ -487,7 +488,7 @@ class SdcProvider:
             xaddrs.append(f'{self._urlschema}://{addr}:{port}/{self.path_prefix}')
         return xaddrs
 
-    def _send_episodic_reports(self, transaction_processor):
+    def _send_episodic_reports(self, transaction_processor: TransactionProcessor):
         mdib_version_group = self._mdib.mdib_version_group
         if transaction_processor.has_descriptor_updates:
             port_type_impl = self.hosted_services.description_event_service
@@ -538,11 +539,12 @@ class SdcProvider:
             port_type_impl = self.hosted_services.waveform_service
             port_type_impl.send_realtime_samples_report(states, mdib_version_group)
 
-    def _send_rt_notifications(self, rt_states):
+    def _send_rt_notifications(self, rt_states: list[AbstractStateProtocol]):
         if len(rt_states) > 0:
             port_type_impl = self.hosted_services.waveform_service
             port_type_impl.send_realtime_samples_report(rt_states, self._mdib.mdib_version_group)
 
-    def set_used_compression(self, *compression_methods):
+    def set_used_compression(self, *compression_methods: str):
+        """Set supported compression methods, e.g. 'gzip'."""
         del self._compression_methods[:]
         self._compression_methods.extend(compression_methods)