+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - change python classes of `addressing_types.py` to match ws-addressing standard of 2006 instead of 2004 
 - The final OperationInvokedReport has OperationTargetRef parameter set. 
   This required refactoring of Operations handling.
+- moved waveform generator from mdib to roles.waveformprovider
 
 ## [2.0.0a6] - 2023-09-11